+++ sample/AppHost/AppHost.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
    <ItemGroup>
-    <ProjectReference Include="..\..\src\Aspire.Temporal.Server\Aspire.Temporal.Server.csproj"  IsAspireProjectResource="false" />
+    <ProjectReference Include="..\..\src\InfinityFlow.Aspire.Temporal\InfinityFlow.Aspire.Temporal.csproj" IsAspireProjectResource="false" />
     <ProjectReference Include="..\Api\Api.csproj" />
     <ProjectReference Include="..\Worker\Worker.csproj" />
   </ItemGroup>
