--- src/Aspire.Temporal.Server/Enums.cs
@@ -1,22 +1,0 @@
-﻿namespace Aspire.Temporal.Server;
-
-public enum LogFormat
-{
-    Json,
-    Pretty
-}
-
-public enum LogLevel
-{
-    Debug,
-    Info,
-    Warn,
-    Error,
-    Fatal
-}
-
-public enum SQLitePragma
-{
-    JournalMode,
-    Synchronous
-}
