--- temporal-server-aspire-component.sln
@@ -1,60 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppHost", "sample\AppHost\AppHost.csproj", "{52057B80-B01E-46C3-96F5-228D0B3BCB48}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceDefaults", "sample\ServiceDefaults\ServiceDefaults.csproj", "{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{0815F1E4-424A-4A05-A12E-B2BDAF2AA142}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{70F8209F-4A68-4F27-AF37-3A76C8EA0642}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aspire.Temporal.Server", "src\Aspire.Temporal.Server\Aspire.Temporal.Server.csproj", "{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api", "Sample\Api\Api.csproj", "{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Worker", "sample\Worker\Worker.csproj", "{8B68A40A-674D-4989-B366-1EE0A633F0D9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{52057B80-B01E-46C3-96F5-228D0B3BCB48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{52057B80-B01E-46C3-96F5-228D0B3BCB48}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{52057B80-B01E-46C3-96F5-228D0B3BCB48}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{52057B80-B01E-46C3-96F5-228D0B3BCB48}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04}.Release|Any CPU.Build.0 = Release|Any CPU
-		{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8B68A40A-674D-4989-B366-1EE0A633F0D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B68A40A-674D-4989-B366-1EE0A633F0D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B68A40A-674D-4989-B366-1EE0A633F0D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B68A40A-674D-4989-B366-1EE0A633F0D9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{52057B80-B01E-46C3-96F5-228D0B3BCB48} = {0815F1E4-424A-4A05-A12E-B2BDAF2AA142}
-		{4B893ACC-7DDF-43CE-95A6-104CA8AEAB04} = {0815F1E4-424A-4A05-A12E-B2BDAF2AA142}
-		{88EE3FBD-5AF8-4F75-A0C7-BC593EC8F90F} = {70F8209F-4A68-4F27-AF37-3A76C8EA0642}
-		{1FFA9DCE-3368-4AFA-960A-0AD2C501CA1B} = {0815F1E4-424A-4A05-A12E-B2BDAF2AA142}
-		{8B68A40A-674D-4989-B366-1EE0A633F0D9} = {0815F1E4-424A-4A05-A12E-B2BDAF2AA142}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C4C2C7A9-980A-4B6D-A624-60A0FBD99DE0}
-	EndGlobalSection
-EndGlobal
